Refactor noise models
